+++ NEWS.md
@@ -1,5 +1,8 @@
 # SeuratIntegrate (development version)
 
+* Add support for scGraph metric (https://doi.org/10.1101/2024.04.02.587824)
+
+
 # SeuratIntegrate 0.4.0
 
 * Initial public release